test: keep discovery parse fixture out of Go linters
