+++ cmd/gocoveragecheck/testdata/functional-discovery/broken_test.txt
